Rename markdown continuation file to MultiParHTMLBlock
